+++ askbot/media/js/askbot/utils/text_property_editor.js
@@ -52,7 +52,7 @@ TextPropertyEditor.prototype.makeEditor = function () {
         me.saveData();
     });
 
-    $(document).append(editor.getElement());
+    $('body').append(editor.getElement());
     return editor;
 };